refactor: adjust code
