+++ src/album/index.ts
@@ -1,0 +1,3 @@
+export * from './album.dto';
+export * from './album.service';
+export * from './album.module';
