+++ src/app.module.ts
@@ -1,13 +1,13 @@
 import { Module } from '@nestjs/common';
-import { UserModule } from './user/user.module';
-import { ArtistModule } from './artist/artist.module';
-import { TrackModule } from './track/track.module';
-import { AlbumModule } from './album/album.module';
-import { FavModule } from './fav/fav.module';
 import { ConfigModule, ConfigService } from '@nestjs/config';
-import { appConfig, dbConfig, authConfig, configSchema } from './config';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AuthModule } from './auth/auth.module';
+import { AuthModule } from './auth';
+import { UserModule } from './user';
+import { ArtistModule } from './artist';
+import { TrackModule } from './track';
+import { AlbumModule } from './album';
+import { FavModule } from './fav';
+import { appConfig, dbConfig, authConfig, configSchema } from './config';
 
 @Module({
   imports: [
