+++ src/artist/artist.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
-import { ArtistController } from './artist.controller';
-import { ArtistService } from './artist.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Artist } from '../lib/entities';
+import { ArtistController } from './artist.controller';
+import { ArtistService } from './artist.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Artist])],
