+++ src/artist/index.ts
@@ -1,0 +1,3 @@
+export * from './artist.dto';
+export * from './artist.service';
+export * from './artist.module';
