+++ src/auth/index.ts
@@ -1,1 +1,1 @@
-export * from './decorators';
+export * from './auth.module';
