+++ src/lib/decorators/auth.decorator.ts
@@ -1,5 +1,5 @@
 import { applyDecorators, UseGuards } from '@nestjs/common';
-import { AccessJwtAuthGuard } from '../guards';
+import { AccessJwtAuthGuard } from '../../auth/guards';
 
 export const Auth = () => {
   return applyDecorators(UseGuards(AccessJwtAuthGuard));
