+++ src/track/track.controller.ts
@@ -10,7 +10,7 @@ import {
   Post,
   Put,
 } from '@nestjs/common';
-import { Auth } from '../auth';
+import { Auth } from '../lib/decorators';
 import { CreateTrackDto, PublicTrackDTO, UpdateTrackDto } from './track.dto';
 import { TrackService } from './track.service';
