+++ src/track/track.dto.ts
@@ -1,6 +1,6 @@
 import { Expose } from 'class-transformer';
 import { IsInt, IsOptional, IsString, IsUUID } from 'class-validator';
-import { BaseDTO } from 'src/utilities/base-dto';
+import { BaseDTO } from '../utilities';
 
 export class CreateTrackDto {
   @IsString()
