+++ src/track/track.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
-import { TrackController } from './track.controller';
-import { TrackService } from './track.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Album, Artist, Track } from '../lib/entities';
+import { TrackController } from './track.controller';
+import { TrackService } from './track.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Track, Artist, Album])],
