+++ src/user/index.ts
@@ -1,2 +1,3 @@
 export * from './user.service';
+export * from './user.dto';
 export * from './user.module';
