+++ src/user/user.dto.ts
@@ -1,6 +1,6 @@
 import { Expose, Type } from 'class-transformer';
 import { IsString } from 'class-validator';
-import { BaseDTO } from 'src/utilities/base-dto';
+import { BaseDTO } from '../utilities';
 
 export class CreateUserDTO {
   @IsString()
